refactor(unit):Refactor the unit type and schema
